requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
beautifulsoup4
requests
streamlit
groq
langchain-groq
faiss-cpu
gunicorn
sentence-transformers 